Submit and remove survey actions added

+++ survey/src/main/java/survey/SurveyRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import survey.Survey;
 
-public interface SurveyRepository extends CrudRepository<Survey, Long> {
+public interface SurveyRepository extends CrudRepository<Survey, Integer> {
 
 }
